cleaverspace_api/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
Django==4.0
django-cors-headers==3.10.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangotoolbox @ git+https://github.com/django-nonrel/djangotoolbox@7e2c47e2ab5ee98d0c0f47c5ccce08e75feda48a
djongo==1.3.6
dnspython==2.1.0
fonttools==4.28.5
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
mongo==0.2.0
mongoengine==0.23.1
numpy==1.21.5
packaging==21.3
Pillow==8.4.0
pymongo==3.12.3
pyparsing==3.0.6
python-dateutil==2.8.2
python-magic==0.4.24
python-magic-bin==0.4.14
python-snappy==0.6.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
urllib3==1.26.7
winkerberos==0.8.0
